+++ src/common/consts.spec.ts
@@ -1,0 +1,5 @@
+import consts from './consts'
+
+it('has default export', () => {
+  expect(consts.defaultMatcher).not.toEqual(undefined)
+})